Issue #4
Overwrote the do_GET, do_POST methods (provided by pythons http.server BaseHTTPRequestHandler library) to do some basic request handling
Note that the BaseHTTPRequestHandler library that's being leveraged here handles a lot of the parsing for us. The only thing we need to need to perform custom parsing on is the body attached to the POST request. Body can currently only support form-data, but has no way of distinguishing form-data from non-form-data.

Note that while I was finishing this up, I noticed that the headers object isn't a dictionary but a HTTPMessage object. This is fine for now, but will need to get looked at later