Update request.py
Fix issue where __rep__ function didn't print headers object

since the headers object is now a HTTPMessage object, we need to create some custom str representation around the object. I don't want to just cast it to a dictionary and lose all the functionality around the HTTPMessage object, so I'm just creating an exception in the __repr__ of the Request object that it feeds in to.